// src/api/model/loanProductGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoanProductGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    name: string;
    rate: number;
    /**
     * Type of the loan
     */
    type: LoanProductGet.TypeEnum;
}
export namespace LoanProductGet {
    export type TypeEnum = 'PersonalLoan' | 'HomeMortgage' | 'CreditCard';
    export const TypeEnum = {
        PersonalLoan: 'PersonalLoan' as TypeEnum,
        HomeMortgage: 'HomeMortgage' as TypeEnum,
        CreditCard: 'CreditCard' as TypeEnum
    };
}